requirements.txt
streamlit
google-generativeai> = 1.2.0
pymongo>=4.6.1
python-dotenv>=1.0.1
certifi


# google-genai
# pymongo>=4.6.1
# python-dotenv>=1.0.1
# certifi
